+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # INLib CHANGELOG
 
-## 3a
+## 3.0
 
 - Added the property `forwardSegueForUnwinding` to INNavigationController and the possibility to forward any call of segueForUnwindingToViewController:fromViewController:identifier: to the destination controller.
 - Removed methods from UIDevice+INExtensions: hasRetinaDisplay, has3Dot5InchesDisplay, has4InchesDisplay